% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-dplyr-tbl.R
\name{write_dbplyr_tbl}
\alias{write_dbplyr_tbl}
\title{Write dbplyr query to snowflake database}
\usage{
write_dbplyr_tbl(con, dplyr_tbl, table_name, schema)
}
\arguments{
\item{con}{connection to database}

\item{table_name}{the name of the table to be written}

\item{schema}{the schema to be used}

\item{dbplyr_tbl}{dbplyr query}
}
\description{
This is a simple function to write a dbplyr query to a demo snowflake database
}
